refactor: move common serializations to one directory
