+++ crates/proof-of-sql/src/base/mod.rs
@@ -15,9 +15,9 @@ pub(crate) mod polynomial;
 /// Module for Proof of SQL datetime types.
 pub mod posql_time;
 pub(crate) mod proof;
-mod standard_binary_serde;
+mod standard_serializations;
 pub use proof::{PlaceholderError, PlaceholderResult};
-pub use standard_binary_serde::{
+pub use standard_serializations::binary::{
     try_standard_binary_deserialization, try_standard_binary_serialization,
 };
 pub(crate) mod ref_into;
