+++ crates/proof-of-sql/src/base/standard_serializations/binary.rs
